BACKPORT: of: make unittest select OF_EARLY_FLATTREE instead of depend on it
The DT unittest currently requires an arch (typically) to select
OF_EARLY_FLATTREE. Remove this dependency by selecting it directly so that
the unittest can be enabled easily on any architecture. With this and the
prior commit, we can easily enable and run unittests starting with x86
defconfig rather than hunting for the combination of config options to
enable OF on x86.

Signed-off-by: Rob Herring <robh@kernel.org>
Cc: Geert Uytterhoeven <geert@linux-m68k.org>
Cc: Pantelis Antoniou <pantelis.antoniou@konsulko.com>
Acked-by: Grant Likely <grant.likely@lianro.org>
(cherry picked from commit 649e0a77e28a7796bf62bfda0fe3f2aee094bd58)
[astrachan: OF_SELFTEST was renamed to OF_UNITTEST upstream, but we
            don't need to make that change]
Bug: 70214720
Change-Id: Ia064a5743e734f5d4436ca651c3169b82a6b68aa
Signed-off-by: Alistair Strachan <astrachan@google.com>