soc: qcom: Fix blackghost SSR handling during TWM mode exit
There were issues in handling of BG ramdump and SSR specially
during TWM exit, as ramdump was being collected during cold
boot as well. This change collect ramdump only if it is
established that we are coming out of TWM mode.

Also irq registration failure during SSR is fixed by this.

Change-Id: Ia07c942f36fc1537b58dd62d4ca6feefbfba3921
Signed-off-by: Avaneesh Kumar Dwivedi <akdwived@codeaurora.org>